requirements.txt
potassium>=0.0.9
torch
transformers
requests
pillow